Extract useRowSelection composable from EditorApp.vue
Move row selection state and logic (~200 lines) into dedicated composable:
- State: selectedRowIndices, lastClickedRowIndex, expandedCommentRowIndex
- Computed: selectedRowIndex, sortedSelectedIndices, canMoveUp, canMoveDown
- Functions: handleRowClick, handleScrollToRow, clearRowSelection, isRowSelected

Zero functional changes - pure refactoring.
