+++ editor/src/components/EditorApp.vue
@@ -43,6 +43,7 @@ import { useMappingCache, type CachedMapping } from '../composables/useMappingCa
 import { useActionHistory, type DeserializationContext } from '../composables/useActionHistory';
 import { useVariableUsage } from '../composables/useVariableUsage';
 import { useSkipAnalysis } from '../composables/useSkipAnalysis';
+import { useRowSelection } from '../composables/useRowSelection';
 import { SetRowColorCommand, SetRowCommentCommand } from '../commands';
 import { MIDI_LEARN_FUNCTION_KEY } from '../constants/midi';
 import { COLOR_PALETTE } from '../constants/colors';
@@ -56,22 +57,22 @@ const { isRowSkipped } = useSkipAnalysis(mappingDocument);
 const currentlySelectedSourceTypes = ref(new Array<MappingType>());
 const currentlySelectedDestinationTypes = ref(new Array<MappingType>());
 
-// Multi-selection state
-const selectedRowIndices = ref<Set<number>>(new Set());
-const lastClickedRowIndex = ref<number | null>(null);
-
-// Comment section expansion state
-const expandedCommentRowIndex = ref<number | null>(null);
-
-// Computed for backwards compatibility with single selection UI
-const selectedRowIndex = computed(() =>
-  selectedRowIndices.value.size === 1
-    ? [...selectedRowIndices.value][0]
-    : null
-);
-
 // Row comments storage (keyed by row index)
 const rowComments = ref<Record<number, string>>({});
+
+// Row selection composable
+const {
+  selectedRowIndices,
+  expandedCommentRowIndex,
+  selectedRowIndex,
+  sortedSelectedIndices,
+  canMoveUp,
+  canMoveDown,
+  handleRowClick,
+  handleScrollToRow,
+  clearRowSelection,
+  isRowSelected
+} = useRowSelection(70, rowComments);
 
 // Row colors storage (keyed by row index)
 const rowColors = ref<Map<number, string>>(new Map());
@@ -458,113 +459,6 @@ async function handleSlotSwitch(slot: 'A' | 'B'): Promise<void> {
   }
 }
 
-// Row selection functions
-function handleRowClick(rowIndex: number, event: MouseEvent): void {
-  if (event.shiftKey && lastClickedRowIndex.value !== null) {
-    // Shift+Click: Select range with sticky comment state
-    const wasCommentOpen = expandedCommentRowIndex.value !== null;
-    const start = Math.min(lastClickedRowIndex.value, rowIndex);
-    const end = Math.max(lastClickedRowIndex.value, rowIndex);
-    for (let i = start; i <= end; i++) {
-      selectedRowIndices.value.add(i);
-    }
-    if (wasCommentOpen) {
-      // Apply comment state to newly selected row
-      if (rowComments.value[rowIndex] === undefined) {
-        rowComments.value[rowIndex] = '';
-      }
-      expandedCommentRowIndex.value = rowIndex;
-    } else {
-      expandedCommentRowIndex.value = null;
-    }
-  } else if (event.ctrlKey || event.metaKey) {
-    // Ctrl/Cmd+Click: Toggle individual with sticky comment state
-    const wasCommentOpen = expandedCommentRowIndex.value !== null;
-    if (selectedRowIndices.value.has(rowIndex)) {
-      selectedRowIndices.value.delete(rowIndex);
-      // Clear comment expansion if this row was expanded
-      if (expandedCommentRowIndex.value === rowIndex) {
-        expandedCommentRowIndex.value = null;
-      }
-    } else {
-      selectedRowIndices.value.add(rowIndex);
-      if (wasCommentOpen) {
-        // Apply comment state to newly selected row
-        if (rowComments.value[rowIndex] === undefined) {
-          rowComments.value[rowIndex] = '';
-        }
-        expandedCommentRowIndex.value = rowIndex;
-      } else {
-        expandedCommentRowIndex.value = null;
-      }
-    }
-  } else {
-    // Normal click: Simple three-click cycle with sticky comment state
-    const isOnlySelected = selectedRowIndices.value.size === 1 && selectedRowIndices.value.has(rowIndex);
-    const isCommentExpanded = expandedCommentRowIndex.value === rowIndex;
-    const wasCommentOpen = expandedCommentRowIndex.value !== null; // Track if ANY comment was open
-
-    if (!selectedRowIndices.value.has(rowIndex)) {
-      // Click 1: Select row and apply current comment state
-      selectedRowIndices.value.clear();
-      selectedRowIndices.value.add(rowIndex);
-      if (wasCommentOpen) {
-        // Previous selection had comment open, so open this one too
-        if (rowComments.value[rowIndex] === undefined) {
-          rowComments.value[rowIndex] = '';
-        }
-        expandedCommentRowIndex.value = rowIndex;
-      } else {
-        // Previous selection had no comment, so don't open this one
-        expandedCommentRowIndex.value = null;
-      }
-    } else if (isOnlySelected && !isCommentExpanded) {
-      // Click 2: Open comment section
-      if (rowComments.value[rowIndex] === undefined) {
-        rowComments.value[rowIndex] = '';
-      }
-      expandedCommentRowIndex.value = rowIndex;
-    } else if (isOnlySelected && isCommentExpanded) {
-      // Click 3: Deselect row
-      selectedRowIndices.value.clear();
-      expandedCommentRowIndex.value = null;
-    } else {
-      // Clicked on a different row in a multi-selection - switch to single selection
-      selectedRowIndices.value.clear();
-      selectedRowIndices.value.add(rowIndex);
-      if (wasCommentOpen) {
-        // Previous selection had comment open, apply to new selection
-        if (rowComments.value[rowIndex] === undefined) {
-          rowComments.value[rowIndex] = '';
-        }
-        expandedCommentRowIndex.value = rowIndex;
-      } else {
-        expandedCommentRowIndex.value = null;
-      }
-    }
-  }
-  lastClickedRowIndex.value = rowIndex;
-  // Force reactivity
-  selectedRowIndices.value = new Set(selectedRowIndices.value);
-}
-
-
-// Handle scroll to row from LogMonitor
-function handleScrollToRow(rowIndex: number): void {
-  // Select the row (no comment opens)
-  selectedRowIndices.value.clear();
-  selectedRowIndices.value.add(rowIndex);
-  expandedCommentRowIndex.value = null;
-  selectedRowIndices.value = new Set(selectedRowIndices.value);
-  
-  // Scroll the row into view
-  const rowElement = document.querySelector(`[data-row-index="${rowIndex}"]`);
-  if (rowElement) {
-    rowElement.scrollIntoView({ behavior: 'smooth', block: 'center' });
-  }
-}
-
-
 // Helper function for MIDI learn source function update
 function handleSourceMidiLearnFunctionUpdate(rowIndex: number, functionKey: number): void {
   const sourceFunc = currentlySelectedSourceTypes.value[rowIndex].functions.find((f: MappingTuple) => f.key === functionKey);
@@ -626,16 +520,6 @@ function handleMidiLearnComplete(currentRowIndex: number): void {
   }
 }
 
-function clearRowSelection(): void {
-  selectedRowIndices.value.clear();
-  selectedRowIndices.value = new Set();
-  lastClickedRowIndex.value = null;
-}
-
-function isRowSelected(rowIndex: number): boolean {
-  return selectedRowIndices.value.has(rowIndex);
-}
-
 // Variable slider handler
 function updateVariableValue(variableIndex: number, value: number): void {
   if (variableIndex >= 0 && variableIndex < mappingDocument.value.variables.length) {
@@ -686,22 +570,6 @@ function formatRowIndex(index: number): string {
   }
   return index.toString();
 }
-
-// Multi-selection computed properties
-const sortedSelectedIndices = computed(() =>
-  [...selectedRowIndices.value].sort((a, b) => a - b)
-);
-
-const canMoveUp = computed(() => {
-  if (selectedRowIndices.value.size === 0) return false;
-  return sortedSelectedIndices.value[0] > 0;
-});
-
-const canMoveDown = computed(() => {
-  if (selectedRowIndices.value.size === 0) return false;
-  const maxIndex = mappingDocument.value.rows.length - 1;
-  return sortedSelectedIndices.value[sortedSelectedIndices.value.length - 1] < maxIndex;
-});
 
 // Multi-selection toolbar handlers
 function handleMultiCut(): void {
